refactor scroller and clean up
